lambda/requirements.txt
boto3==1.9.216
ask-sdk-core==1.11.0
IMDbPY==2022.7.9
Requests==2.31.0